+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - **_Breaking Change_**: `ActionExec` will take `ActionInput` as the input. [#6](https://github.com/DiscreteTom/retsac/issues/6)
     - Rename `ActionAcceptedOutput` to `AcceptedActionOutput`, and make it a class, rewrite its logics.
     - `Action.mute/check/reject/then` will take `AcceptedActionOutput` as the callback's param.
+    - Regex should not starts with `^` and will auto have sticky flag.
+    - Export `Action.simple/match`. Add `options.autoSticky/rejectCaret` for `Action.match`.
   - Feat: `SimpleActionExec` can accept `string/SimpleAcceptedActionOutput` as the output.
 - Optimize performance & comments.